requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
cffi==1.15.0
cryptography==36.0.1; python_version >= '3.6'
dj-database-url==0.5.0
django-filter==21.1
django-rest-knox==4.1.0
django==3.2.11
djangorestframework==3.13.1
drf-nested-routers==0.93.4
gunicorn==20.1.0
psycopg2==2.9.3
pycparser==2.21
pytz==2021.3
setuptools==60.5.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
whitenoise==5.3.0
